auth messages confirmation

+++ src/app/components/user/register/register.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { UserService } from 'src/app/services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-register',
@@ -19,7 +20,7 @@ export class RegisterComponent implements OnInit {
     this.signUpForm = this.formBuilder.group({
       name: ['', [Validators.required]],
       email: ['', [Validators.required, Validators.email]],
-      password: ['', [Validators.required]],
+      password: ['', [Validators.required,Validators.minLength(5)]],
     });
    }
 
@@ -27,23 +28,43 @@ export class RegisterComponent implements OnInit {
   }
   get f() { return this.signUpForm.controls; }
 
-  signUp() {
+  signUp():void {
     this.submitted = true;
     console.log(this.signUpForm.value);
-    
-    this.userService.signUp(this.signUpForm.value).subscribe(
-      (res: any) => {
-        console.log(res);
-        if(res.status == 'success'){
-        console.log(res.authorisation.token);
-        localStorage.setItem('access_token', res.authorisation.token);
-        this.router.navigate(['/movies']);
+    if( this.signUpForm.invalid ) return;
+    Swal.fire({
+      title: 'Loading ...',
+      timerProgressBar: true,
+      didOpen: () => {
+        Swal.showLoading()
+      
+        this.userService.signUp(this.signUpForm.value).subscribe(
+          (res: any) => {
+            console.log(res);
+            Swal.close();
+            if(res.status == 'success'){
+            console.log(res.authorisation.token);
+            localStorage.setItem('access_token', res.authorisation.token);
+            this.router.navigate(['/movies']);
+            } else if (res.status == 'failed') {
+              for ( const key in res.message ) {
+                this.signUpForm.get(key)?.setErrors( { server: res.message[key] } )
+              }
+          }
+            
+          }, (err: any) => {
+            Swal.close();
+            console.log(err);
+            Swal.fire({
+              position: 'center',
+              icon: 'error',
+              title: 'Something is error please try again later',
+              showConfirmButton: true,
+              // timer: 2000
+            })
+          }
+        )
       }
-        
-      }, (err: any) => {
-        console.log(err);
-        
-      }
-    )
+    })
   }
 }
